Fix build with Bazel
When building Rust applications with Bazel instead of cargo, typically a
sandbox is in place, where each process invocation (rustc, build.rs
program, etc.) runs in its own "root", where the view on the filesystem
is the same, but the absolute paths are different. As a consequence,
CARGO_MANIFEST_DIR differs between the invocation of build.rs and rustc.

That means that absolute paths in the generated webgl_exts.rs don't
work. Fortunately, the source files are relative to the
CARGO_MANIFEST_DIR, which, while changing between invocations, remains
valid.