Remove deprecated `webgl` support via `stdweb`
It appears that the only "target" our `webgl_generator` supports is
`stdweb`, a project that hasn't been maintained and updated either.
Remove everything related to it to simplify maintenance and progress of
this crate.

As already mentioned in #524 these bindings appear to be fully superseded by
other projects.

The `webgl` XML contents remain as part of `khronos_api`.