khronos_api: Resolve WebGL extension files relative to `CARGO_MANIFEST_DIR`
Fixes #536
Fixes #549
Fixes #554

There are not one, not two but three PRs attempting to fix Bazel build
support, which seems to containerize or sandbox the build script such
that the paths it uses to find the crate source (e.g. `current_dir()`)
and XML files within is different than when the generated source is
being compiled, leading to files no longer being found.

The most simple solution of course is to remove all assumptions and
references to the working directory or absolute location of the crate
when the build script is being ran.  This can only be used to find the
files, after which their relative paths (**within** the crate source)
will be emitted to the output file with a `concat!()` directive that
prepends it with `env!("CARGO_MANIFEST_DIR")` whenever the source is
finally being compiled (in a different sandbox).