gl_generator: Make lookups more robust based on unstripped name
By stripping and later appending a *possible* prefix for enum constants
and command names we're not only spending a lot of CPU and human
cycles (because this logic is impossible to understand), we're also
loosing out on a few WGL commands that weren't prefixed with the
`wgl` name.  Conveniently these functions were referencing a typo'd
`PIXELFORMATDESCRIPTOR` without leading `P`, which might otherwise not
have been found.

Do this by keeping all commands and enums keyed by their full name
in `BTreeMap`s so that we can also track completely accurately when a
`<delete>` or `<require>` would erroneously specify a name that doesn't
exist.